fix: standalone HEC Event writer (#380)
When create_from_token() is invoked, it triggers the HECEventWriter
constructor with scheme, host, and port set to None. This leads to a
get_splunkd_access_info() call, which throws an exception if the HEC URI
doesn’t reference the local machine. As this call is unnecessary in such
cases, its invocation has been relocated within the if-statement